feat: headless functionality in hook
