+++ src/index.js
@@ -1,3 +1,8 @@
 import OAuth2Login from './OAuth2Login';
+import useOAuth2Login from './useOAuth2Login';
+
+export {
+  useOAuth2Login,
+};
 
 export default OAuth2Login;
